# [crypto] Optimize the FromRLP with sync pool