@rlan35 2021-03-29
awesome! Can't wait to see the drop of CPU usage on mainnet.