@JackyWYX 2021-03-29
> awesome! Can't wait to see the drop of CPU usage on mainnet.

Well, this is not going to help with CPU issue a lot as far as I see. It's just a very small fix from my previous experience on Golang. ;)